alt slice syntax transforms
ASL contains a syntactic ambiguity between the use of < and > as comparision
operators and their use as bitslice delimiters.  For example:

     F(a < b, c, d, e)  // four function arguments, first performs comparision

and
     F(a<b,c,d,e>) // a single function argument, argument performs bitslice

As an experiment (that may go away in the future) I have added an optional
transform to the ASL extraction code that converts the <> bitslice syntax
to use [] instead.  With this change, the examples become:

     F(a < b, c, d, e) // unchanged

and
     F(a[b,c,d,e]) // use [] for bitslices

This also has ambiguities (with array indexing and function calls) but they are
not as severe as the ambiguities caused by the <> syntax.