Fix handling of instruction encoding that use 'psbits'
Some groups of instructions share a piece of common ASL code and
differ only in one or two bits of their encoding.
For example, MOVZ and MOVK or ADD and SUB.

The XML files contain copies of the same code and slightly different encoding
diagrams for each individual instruction but the instrs2asl.py script extracts
just one copy of the code and just one encoding diagram.

To do this correctly, instrs2asl.py has to identify which bits are common to
all instructions and which bits are used to distinguish the different
instructions in the group.  This is easy to do because the XML marks those
fields of the encoding with the property "psbits".  This patch implements
that check.